requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
xlsxwriter==3.1.0
requests==2.31.0
apscheduler==3.10.1
numpy==1.25.2
python-magic==0.4.27
gunicorn==21.2.0
flask-wtf==1.1.1
pytest==7.4.0
